pyproject.toml
[tool.poetry]
name = "seo_friendly_urls_django"
version = "0.0.1"
description = ""
authors = ["Dani Hodovic <dani.hodovic@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"

[tool.poetry.dev-dependencies]
isort = {version = "^4.3", allows-prereleases = true}
black = {version = "^18.3-alpha.0", allows-prereleases = true}
pudb = "^2019.1"
django-extensions = "^2.2"
werkzeug = "^0.15.5"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
